fix missing device assignment
