+++ app/src/main/java/com/framgia/fdms/screen/request/requestmanager/RequestManagerViewModel.java
@@ -21,11 +21,11 @@ import com.framgia.fdms.data.model.Status;
 import com.framgia.fdms.data.model.User;
 import com.framgia.fdms.screen.assignment.AssignmentActivity;
 import com.framgia.fdms.screen.assignment.AssignmentType;
-import com.framgia.fdms.screen.requestcreation.RequestCreationActivity;
-import com.framgia.fdms.screen.selection.SelectionActivity;
 import com.framgia.fdms.screen.request.OnRequestClickListenner;
 import com.framgia.fdms.screen.request.userrequest.UserRequestAdapter;
+import com.framgia.fdms.screen.requestcreation.RequestCreationActivity;
 import com.framgia.fdms.screen.requestdetail.RequestDetailActivity;
+import com.framgia.fdms.screen.selection.SelectionActivity;
 import com.framgia.fdms.utils.navigator.Navigator;
 import java.util.ArrayList;
 import java.util.List;
@@ -259,7 +259,7 @@ public class RequestManagerViewModel extends BaseFragmentModel
 
     @Override
     public void onDetailRequestClick(Request request) {
-        mFragment.startActivityForResult(RequestDetailActivity.newInstance(mContext, request),
+        mFragment.startActivityForResult(RequestDetailActivity.getInstance(mContext, request),
             REQUEST_DETAIL);
     }
